pdf-to-md/requirements.txt
pymupdf>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
selenium>=4.15.0
pdf2image>=1.16.0
